# perf: optimize curve membership test