--- a/ecc/bls24-317/g2.go
+++ b/ecc/bls24-317/g2.go
@@ -398,8 +398,7 @@ func (p *G2Jac) IsOnCurve() bool {
 	left.Square(&p.Y)
 	right.Square(&p.X).Mul(&right, &p.X)
 	ZZ.Square(&p.Z)
-	tmp.Square(&ZZ).
-		Mul(&tmp, &ZZ)
+	tmp.Square(&ZZ).Mul(&tmp, &ZZ)
 	tmp.MulBybTwistCurveCoeff(&tmp)
 	right.Add(&right, &tmp)
 	return left.Equal(&right)