--- a/ecc/bw6-633/pairing.go
+++ b/ecc/bw6-633/pairing.go
@@ -320,7 +320,7 @@ func (p *g2Proj) doubleStep(evaluations *lineEvaluation) {
 	C.Square(&p.z)
 	D.Double(&C).
 		Add(&D, &C)
-	E.Mul(&D, &bTwistCurveCoeff)
+	E.Double(&D).Double(&E).Double(&E)
 	F.Double(&E).
 		Add(&F, &E)
 	G.Add(&B, &F)