--- a/ecc/bw6-761/g1.go
+++ b/ecc/bw6-761/g1.go
@@ -406,8 +406,8 @@ func (p *G1Jac) IsOnCurve() bool {
 	left.Square(&p.Y)
 	right.Square(&p.X).Mul(&right, &p.X)
 	ZZ.Square(&p.Z)
-	tmp.Square(&ZZ).
-		Mul(&tmp, &ZZ)
+	tmp.Square(&ZZ).Mul(&tmp, &ZZ)
+	// Mul tmp by bCurveCoeff=-1
 	tmp.Neg(&tmp)
 	right.Add(&right, &tmp)
 	return left.Equal(&right)