--- a/internal/generator/ecc/template/point.go.tmpl
+++ b/internal/generator/ecc/template/point.go.tmpl
@@ -434,13 +434,12 @@ func (p *{{ $TJacobian }}) FromAffine(Q *{{ $TAffine }}) *{{ $TJacobian }} {
 
 // IsOnCurve returns true if p in on the curve
 func (p *{{ $TJacobian }}) IsOnCurve() bool {
-	var left, right, tmp  {{.CoordType}}
+	var left, right, tmp, ZZ  {{.CoordType}}
 	left.Square(&p.Y)
 	right.Square(&p.X).Mul(&right, &p.X)
-	tmp.Square(&p.Z).
-		Square(&tmp).
-		Mul(&tmp, &p.Z).
-		Mul(&tmp, &p.Z).
+	ZZ.Square(&p.Z)
+    tmp.Square(&ZZ).
+		Mul(&tmp, &ZZ).
 		{{- if eq .PointName "g1"}}
 			Mul(&tmp, &bCurveCoeff)
 		{{- else}}