--- a/internal/generator/ecc/template/point.go.tmpl
+++ b/internal/generator/ecc/template/point.go.tmpl
@@ -439,11 +439,28 @@ func (p *{{ $TJacobian }}) IsOnCurve() bool {
 	right.Square(&p.X).Mul(&right, &p.X)
 	ZZ.Square(&p.Z)
     tmp.Square(&ZZ).
-		Mul(&tmp, &ZZ).
+		Mul(&tmp, &ZZ)
 		{{- if eq .PointName "g1"}}
-			Mul(&tmp, &bCurveCoeff)
+            {{- if or (eq .Name "bls12-381") (eq .Name "bls24-317") (eq .Name "bw6-633")}}
+                tmp.Double(&tmp).Double(&tmp)
+            {{- else if eq .Name "bn254"}}
+                fp.MulBy3(&tmp)
+            {{- else if eq .Name "bw6-761"}}
+                tmp.Neg(&tmp)
+            {{- else if or (eq .Name "bls12-377") (eq .Name "bls12-378") (eq .Name "bls24-315") (eq .Name "bw6-756")}}
+            {{- else}}
+                tmp.Mul(&tmp, &bCurveCoeff)
+            {{- end}}
 		{{- else}}
-			Mul(&tmp, &bTwistCurveCoeff)
+            {{- if or (eq .Name "bls12-377") (eq .Name "bls12-381") (eq .Name "bn254") (eq .Name "bls24-315") (eq .Name "bls24-317")}}
+                tmp.MulBybTwistCurveCoeff(&tmp)
+            {{- else if eq .Name "bw6-761"}}
+                tmp.Double(&tmp).Double(&tmp)
+            {{- else if eq .Name "bw6-633"}}
+                tmp.Double(&tmp).Double(&tmp).Double(&tmp)
+            {{- else}}
+                tmp.Mul(&tmp, &bTwistCurveCoeff)
+            {{- end}}
 		{{- end}}
 	right.Add(&right, &tmp)
 	return left.Equal(&right)