--- a/internal/generator/ecc/template/point.go.tmpl
+++ b/internal/generator/ecc/template/point.go.tmpl
@@ -438,16 +438,19 @@ func (p *{{ $TJacobian }}) IsOnCurve() bool {
 	left.Square(&p.Y)
 	right.Square(&p.X).Mul(&right, &p.X)
 	ZZ.Square(&p.Z)
-    tmp.Square(&ZZ).
-		Mul(&tmp, &ZZ)
+    tmp.Square(&ZZ).Mul(&tmp, &ZZ)
 		{{- if eq .PointName "g1"}}
             {{- if or (eq .Name "bls12-381") (eq .Name "bls24-317") (eq .Name "bw6-633")}}
+                // Mul tmp by bCurveCoeff=4
                 tmp.Double(&tmp).Double(&tmp)
             {{- else if eq .Name "bn254"}}
+                // Mul tmp by bCurveCoeff=3
                 fp.MulBy3(&tmp)
             {{- else if eq .Name "bw6-761"}}
+                // Mul tmp by bCurveCoeff=-1
                 tmp.Neg(&tmp)
             {{- else if or (eq .Name "bls12-377") (eq .Name "bls12-378") (eq .Name "bls24-315") (eq .Name "bw6-756")}}
+                // Mul tmp by bCurveCoeff=1 (nothing to do)
             {{- else}}
                 tmp.Mul(&tmp, &bCurveCoeff)
             {{- end}}
@@ -455,8 +458,10 @@ func (p *{{ $TJacobian }}) IsOnCurve() bool {
             {{- if or (eq .Name "bls12-377") (eq .Name "bls12-381") (eq .Name "bn254") (eq .Name "bls24-315") (eq .Name "bls24-317")}}
                 tmp.MulBybTwistCurveCoeff(&tmp)
             {{- else if eq .Name "bw6-761"}}
+                // Mul tmp by bTwistCurveCoeff=4
                 tmp.Double(&tmp).Double(&tmp)
             {{- else if eq .Name "bw6-633"}}
+                // Mul tmp by bTwistCurveCoeff=8
                 tmp.Double(&tmp).Double(&tmp).Double(&tmp)
             {{- else}}
                 tmp.Mul(&tmp, &bTwistCurveCoeff)