# Description

This PR optimizes the curve membership test `IsOnCurve()` for all curves on both G1 and G2, by:
- [x] saving 1 Fp mul
- [x] specializing the multiplication by the curve coefficient `b`


## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

Old `TestG1AffineIsOnCurve` and `TestG2AffineIsOnCurve`

# How has this been benchmarked?

e.g. BLS12-381 on AWS z1d.large:

`G1`  335 ns vs. **190.6** ns
`G2` 624 ns vs. **495.3** ns

Compared to [geth/kilic](https://github.com/ethereum/go-ethereum/tree/master/crypto/bls12381) (cc @asanso, @MariusVanDerWijden):
`G1` 320.7 ns vs. **190.6** ns
`G2` 917.9 ns vs. **495.3** ns

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules